pyproject.toml
[project]
name = "fast-repl"
version = "0.1.0"
description = "Server to check Lean proofs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "loguru>=0.7.3",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pyright>=1.1.402",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "uvicorn>=0.34.3",
]

[tool.uv]
dev-dependencies = [
    "pytest-cov>=6.2.1",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
addopts = "--asyncio-mode=auto --maxfail=1 --cov=fast_repl "

[tool.mypy]
strict = true
files = ["src"]
ignore_missing_imports = true 

# TODO: Think about what it means to have this lib packaged: Docker vs. pypi vs. cli?
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
